all passed 1

+++ 4-module/2-task/server.js
@@ -3,18 +3,21 @@ const http = require('http');
 const path = require('path');
 const fs = require('fs');
 const { finished } = require('stream');
+const LimitSizeStream = require('./LimitSizeStream');
+const LimitExceededError = require('./LimitExceededError');
 
 const server = new http.Server();
 
 server.on('request', (req, res) => {
   //let file = '';
   
-  function SendError(err, file) {
+  function SendError(err, errCode, file) {
     console.log('MAX: SendError ' + err + ' --- ' + file.path);
 
     if (file) {
       file.once('close', () => {
-        fs.unlinkSync(file.path);    
+        if (fs.existsSync(file.path)) 
+          fs.unlinkSync(file.path);    
         console.log('MAX: SendError unlink');  
       })
 
@@ -22,8 +25,8 @@ server.on('request', (req, res) => {
       console.log('MAX: SendError close');
     }
 
-    res.statusCode = 500;
-    res.end(`Server streams Error: ${err.message}`); 
+    res.statusCode = errCode;
+    res.end(`Server Error: ${err.message}`); 
   }
 
   const pathname = url.parse(req.url).pathname.slice(1);
@@ -46,14 +49,20 @@ server.on('request', (req, res) => {
           break;
         }        
 
-      const file = fs.createWriteStream(filepath);
-      file.on('error', SendError);
       //req.on('error', SendError);
 
+      const file = fs.createWriteStream(filepath);
+      //file.on('error', SendError);
+      
+      const limitStream = new LimitSizeStream({limit: 1*1024*1024});
+        limitStream.on('error', (err) => {
+          SendError(err, 413, file);
+      });
+      
       finished(req, (err) => {
         if (err) {
           console.log('MAX:finished: Stream error');
-          SendError(err, file);
+          SendError(err, 500, file);
         } else {
           console.log('MAX:finished: Stream is done reading.');
           file.close;
@@ -64,7 +73,7 @@ server.on('request', (req, res) => {
         console.log('MAX:finished: Resumed.');
       });   
 
-      req.pipe(file);
+      req.pipe(limitStream).pipe(file);
       
       /*file.once('finish', () => {  
         console.log('MAX:finish EVENT ' + file);
@@ -81,6 +90,6 @@ server.on('request', (req, res) => {
   }
 });
 
-server.on('error', error => {console.error(error());});
+//server.on('error', error => {console.error(error());});
 
 module.exports = server;
